Move actions to separate files
