+++ src/bots/events/IBaseContext.ts
@@ -1,0 +1,7 @@
+export default abstract class IBaseContext {
+  public readonly botName: string;
+
+  protected constructor(botName: string) {
+    this.botName = botName;
+  }
+}
